Move HTTP code in helper.py to a separate file
