--- demo/EV3D4/EV3D4WebControl
@@ -1,13 +1,0 @@
-#!/usr/bin/env python3
-
-from EV3D4 import EV3D4WebControlled
-import logging
-
-logging.basicConfig(level=logging.DEBUG,
-                    format='%(asctime)s %(levelname)5s: %(message)s')
-log = logging.getLogger(__name__)
-
-log.info("Starting EV3D4")
-ev3d4 = EV3D4WebControlled()
-ev3d4.main()  # start the web server
-log.info("Exiting EV3D4")
